requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
billiard==3.6.3.0
celery==4.3.0
Django==3.1
djangorestframework==3.11.0
kombu==4.6.11
pytz==2020.1
redis==3.2.0
sqlparse==0.3.1
vine==1.3.0
